+++ demo/demo_form.php
@@ -5,6 +5,7 @@ $form_def=array(
     'desc'	=>"Please enter a name",
     'type'	=>"text",
     'req'	=>true,
+    'max_length' => 10,
     'html_attributes'   =>array("style"=>"border: 2px solid black;"),
     'check'	=> array("fun", array("js"=>"name_check", "php"=>"name_check")),
   ),
@@ -56,6 +57,7 @@ EOT
     'values_func'=>array("js"=>"fav_hobby_list", "php"=>"fav_hobby_list"),
     'values_mode'=>"values",
     'name'      =>array("en"=>"Favorite hobby", "de"=>"Lieblingshobby"),
+    'include_data' => 'not_null',
   ),
   'birthday'	=>array(
     'name'	=>"Birthday",