+++ test/load.php
@@ -1,0 +1,3 @@
+<?php
+include "modulekit/loader.php"; /* loads all php-includes */
+call_hooks("init");